fix(openai): Responses API - ensure consistent message ID between stream & invoke
